refactor(library/Parser): extracts `ParsingError` into individual library
